traceset_v2: Moved enum to enums package
